# API-contract.yaml
openapi: 3.0.3
info:
  title: Function Storage API
  version: 1.0.0
  description: |
    Единый унифицированный контракт API для работы с табулированными функциями.
    Поддерживает:
      - операции CRUD и поиск по владельцу;
      - арифметические операции над функциями (сложение, вычитание, умножение, деление);
      - численное дифференцирование табулированных функций;
      - импорт/экспорт в текстовом и бинарном формате;
      - сериализацию/десериализацию функций.
servers:
  - url: http://localhost:8080/api
    description: |
      Local development (manual: Tomcat, framework: Spring Boot)
security:
  - BasicAuth: []
paths:
...
  /search/functions/by-name:
    get:
      summary: Найти функции, имя которых содержит подстроку
      tags: [ Search ]
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Подстрока для поиска в имени функции
      responses:
        '200':
          description: Список функций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FunctionListResponse'
...
  /functions:
    post:
      summary: Создать новую табулированную функцию
      tags: [Functions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionCreateRequest'
      responses:
        '201':
          description: Функция создана
          headers:
            Location:
              schema:
                type: string
                example: /api/functions/123
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
  /functions/{id}:
    get:
      summary: Получить функцию по ID (только владельцу)
      tags: [Functions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Функция найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Функция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Обновить имя и/или точки функции (только владельцу)
      tags: [Functions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionUpdateRequest'
      responses:
        '200':
          description: Функция обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Удалить функцию (только владельцу)
      tags: [Functions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Успешно удалено
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /functions/owner/{ownerId}:
    get:
      summary: Получить все функции пользователя
      tags: [Functions]
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список функций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FunctionResponse'
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    FunctionListResponse: # Соответствует TabulatedFunctionListDTO.java
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID функции
        name:
          type: string
          description: Имя функции
        createdAt:
          type: string
          format: date-time
          description: Дата создания функции
        userId:
          type: integer
          format: int64
          description: ID владельца
        username:
          type: string
          description: Имя владельца
        pointsCount:
          type: integer
          minimum: 0
          description: Количество точек в функции
...
    FunctionCreateRequest:
      type: object
      required: [ownerId, name, points]
      properties:
        ownerId:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 100
          example: "my_quadratic"
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointData'
    FunctionUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: "renamed"
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointData'
    FunctionResponse:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointData'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PointData:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
          format: double
          example: 1.0
        y:
          type: number
          format: double
          example: 2.0
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "ValidationFailed"
        message:
          type: string
          example: "Поле 'name' не может быть пустым"
        field:
          type: string
          example: "name"
        timestamp:
          type: string
          format: date-time